+++ Sources/KeyboardKit/_Deprecated/KeyboardLayoutItem+Deprecated.swift
@@ -5,3 +5,9 @@ public typealias KeyboardLayoutItemRow = KeyboardLayoutItem.Row
 
 @available(*, deprecated, renamed: "KeyboardLayoutItem.Rows")
 public typealias KeyboardLayoutItemRows = KeyboardLayoutItem.Rows
+
+@available(*, deprecated, renamed: "KeyboardLayoutItem.Size")
+public typealias KeyboardLayoutItemSize = KeyboardLayoutItem.Size
+
+@available(*, deprecated, renamed: "KeyboardLayoutItem.Width")
+public typealias KeyboardLayoutItemWidth = KeyboardLayoutItem.Width