+++ Tests/KeyboardKitTests/Layout/KeyboardLayoutTests.swift
@@ -12,8 +12,8 @@ import XCTest
 
 class KeyboardLayoutTests: XCTestCase {
 
-    func item(_ width: KeyboardLayoutItemWidth) -> KeyboardLayoutItem {
-        let size = KeyboardLayoutItemSize(width: width, height: 0)
+    func item(_ width: KeyboardLayoutItem.Width) -> KeyboardLayoutItem {
+        let size = KeyboardLayoutItem.Size(width: width, height: 0)
         let item = KeyboardLayoutItem(action: .none, size: size, insets: .init())
         return item
     }